+++ src/h3/__init__.py
@@ -1,8 +1,11 @@
 # flake8: noqa
 
-from .api.set_str import *
-from .api import set_str as h3
+from .api.py_str import *
 from ._version import __version__
+
+#todo: remove in 4.0; only here for backward compatibility
+from .api import py_str as h3
+
 
 from ._cy import (
     H3ValueError,
