+++ src/h3/api/numpy_int.py
