+++ src/h3/api/py_str.py
